remove_apache_http_client - some test migration
